requirements.txt
flask==2.3.2
requests==2.31.0
pytest==7.4.2
pytest-flask==1.2.0
flask-restful==0.3.9
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
python-dotenv==1.0.0